requirements.txt
asgiref==3.4.1
Django==3.2.10
django-ckeditor==6.2.0
django-ckeditor5==0.0.4
django-js-asset==1.2.2
Pillow==8.4.0
PyMySQL==1.0.2
pytz==2021.3
simpleui==4.0.2
sqlparse==0.4.2
typing_extensions==4.0.1
uWSGI==2.0.20
asgiref==3.4.1
Django==3.2.10
django-ckeditor==6.2.0
django-ckeditor5==0.0.4
django-js-asset==1.2.2
extras==1.0.0
Pillow==8.4.0
PyMySQL==1.0.2
pytz==2021.3
simpleui==4.0.2
sqlparse==0.4.2
typing_extensions==4.0.1
uWSGI==2.0.20
